Catching the exceptions if MA or SS cannot be saved

+++ scripts/bellcom_borgerdk.cron.php
@@ -9,9 +9,6 @@ print('==========================' . PHP_EOL);
 print('Started bellcom_borgerdk.cron.php' . PHP_EOL);
 print('==========================' . PHP_EOL);
 
-$queue = DrupalQueue::get(BELLCOM_BORGERDK_QUEUE);
-$queue->deleteQueue();
-
 print('Filling the queue' . PHP_EOL);
 bellcom_borgerdk_cron();
 
@@ -19,19 +16,21 @@ $queue = DrupalQueue::get(BELLCOM_BORGERDK_QUEUE);
 print('Total number of items in queue: ' . $queue->numberOfItems() * BELLCOM_BORGERDK_QUEUE_CHUNK_SIZE . PHP_EOL);
 $total_chunks = $queue->numberOfItems();
 $current_item = 1;
-while ($queue->numberOfItems() > 0) {
+while ($queue->numberOfItems() > 0 && $current_item <= $queue->numberOfItems()) {
   print('Processing chunk : ' . $current_item . '/' . $total_chunks . PHP_EOL);
   $chunk = $queue->claimItem();
-  $items = $chunk->data;
-  if (is_array($items)) {
-    try {
-      print('Article IDs to process: ' . implode(', ', array_keys($chunk->data)) . PHP_EOL);
-      bellcom_borgerdk_queue_worker($chunk->data);
-    } catch (Exception $e){
-      print('Error! Ignoring chunk. Exception: ' . $e->getMessage());
+  if ($chunk) {
+    $items = $chunk->data;
+    if (is_array($items)) {
+      try {
+        print('Article IDs to process: ' . implode(', ', array_keys($chunk->data)) . PHP_EOL);
+        bellcom_borgerdk_queue_worker($chunk->data);
+      } catch (Exception $e){
+        print('Error! Ignoring chunk. Exception: ' . $e->getMessage());
+      }
     }
+    $queue->deleteItem($chunk);
   }
-  $queue->deleteItem($chunk);
   $current_item++;
   print('==========================' . PHP_EOL);
 }
